이지선/4주차/tdd-study/src/main/java/com/gdsc/tddstudy/controller/CalculatorApiController.java
package com.gdsc.tddstudy.controller;

import com.gdsc.tddstudy.component.Calculator;
import com.gdsc.tddstudy.dto.Req;
import com.gdsc.tddstudy.dto.Res;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.*;


@RestController
@RequestMapping("/api")
@RequiredArgsConstructor
public class CalculatorApiController {

    private final Calculator calculator;

    @GetMapping("/sum")
    public int sum(@RequestParam int x, @RequestParam int y) {
        return calculator.sum(x, y);
    }

    @PostMapping("/minus")
    public Res minus(@RequestBody Req req) {
        int result = calculator.mius(req.getX(), req.getY());

        Res res = new Res();
        res.setResult(result);
        res.setResponse(new Res.Body());
        return res;
    }
}